refactor: separated Node class & renamed custom List class - Also added XML documentation comments
