+++ OrganiTask/OrganiTask.csproj
@@ -121,7 +121,8 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Util\LinkedList.cs" />
+    <Compile Include="Util\Collections\Node.cs" />
+    <Compile Include="Util\Collections\OrganiList.cs" />
     <Compile Include="Util\OrganiTaskDB.cs" />
     <Compile Include="Services\AuthService.cs" />
     <EmbeddedResource Include="Forms\Login.resx">
